Implement stacked-histogram-chart and DRY
